CORDA-649: Make default properties to be flexible about various types
