+++ node/src/main/kotlin/net/corda/node/services/messaging/ArtemisMessagingServer.kt
@@ -391,7 +391,7 @@ class ArtemisMessagingServer(override val config: NodeConfiguration,
             // We keep trying until the network map deems the node unreachable and tells us it's been removed at which
             // point we destroy the bridge
             retryInterval = LongPropertiesRegistry.BRIDGE_RETRY_INTERVAL_MS.getOrDefault()
-            retryIntervalMultiplier = 1.5  // Exponential backoff
+            retryIntervalMultiplier = DoublePropertiesRegistry.BRIDGE_RETRY_INTERVAL_MULTIPLIER.getOrDefault()  // Exponential backoff
             maxRetryInterval = LongPropertiesRegistry.BRIDGE_MAX_RETRY_INTERVAL_MS.getOrDefault()
             // As a peer of the target node we must connect to it using the peer user. Actual authentication is done using
             // our TLS certificate.
